Move common test utilities to com.github.behaim.utils
